move db schema file to separate files
